+++ common/envs/constants.ts
@@ -3,6 +3,7 @@ import { DEV_CONFIG } from './dev'
 import { EnvConfig, PROD_CONFIG } from './prod'
 import { THEOREMONE_CONFIG } from './theoremone'
 
+export const BACKGROUND_COLOR = 'bg-gray-50'
 export const ENV = process.env.NEXT_PUBLIC_FIREBASE_ENV ?? 'PROD'
 
 const CONFIGS: { [env: string]: EnvConfig } = {
