+++ web/components/multi-user-transaction-link.tsx
@@ -17,39 +17,24 @@ export type MultiUserLinkInfo = {
 export function MultiUserTransactionLink(props: {
   userInfos: MultiUserLinkInfo[]
   modalLabel: string
+  text: string
 }) {
-  const { userInfos, modalLabel } = props
+  const { userInfos, modalLabel, text } = props
   const [open, setOpen] = useState(false)
   const maxShowCount = 5
   return (
-    <Row>
-      <Button
-        size={'xs'}
-        color={'gray-white'}
-        className={'z-10 mr-1 gap-1 bg-transparent'}
+    <span>
+      <button
+        className={
+          'bg-transparent font-semibold transition-colors hover:text-indigo-500'
+        }
         onClick={(e) => {
           e.stopPropagation()
           setOpen(true)
         }}
       >
-        <Row className={'items-center gap-1 sm:gap-2'}>
-          {userInfos.map(
-            (userInfo, index) =>
-              index < maxShowCount && (
-                <Avatar
-                  username={userInfo.username}
-                  size={'sm'}
-                  avatarUrl={userInfo.avatarUrl}
-                  noLink={userInfos.length > 1}
-                  key={userInfo.username + 'avatar'}
-                />
-              )
-          )}
-          {userInfos.length > maxShowCount && (
-            <span>& {userInfos.length - maxShowCount} more</span>
-          )}
-        </Row>
-      </Button>
+        {text}
+      </button>
       <Modal open={open} setOpen={setOpen} size={'sm'}>
         <Col className="items-start gap-4 rounded-md bg-white p-6">
           <span className={'text-xl'}>{modalLabel}</span>
@@ -70,6 +55,6 @@ export function MultiUserTransactionLink(props: {
           ))}
         </Col>
       </Modal>
-    </Row>
+    </span>
   )
 }
